Refactor specUtils.js
